Extract all the command and rearrange packages
