--- src/main/java/commands/AddRecordCommand.java
@@ -1,4 +1,0 @@
-package commands;
-
-public class AddRecordCommand {
-}
