--- src/main/java/commands/DeletePersonCommand.java
@@ -1,4 +1,0 @@
-package commands;
-
-public class DeletePersonCommand {
-}
