--- src/main/java/commands/DeleteRecordCommand.java
@@ -1,4 +1,0 @@
-package commands;
-
-public class DeleteRecordCommand {
-}
