--- src/main/java/commands/ExitCommand.java
@@ -1,4 +1,0 @@
-package commands;
-
-public class ExitCommand {
-}
