--- src/main/java/commands/ListCommand.java
@@ -1,4 +1,0 @@
-package commands;
-
-public class ListCommand {
-}
