--- src/main/java/commands/RemainCommand.java
@@ -1,4 +1,0 @@
-package commands;
-
-public class RemainCommand {
-}
