+++ src/main/java/seedu/planitarium/exceptions/InvalidIndexException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package seedu.planitarium.exceptions;
 
 /**
  * Thrown to indicate that a string cannot be converted into a valid integer index or index is out of bounds.
