+++ src/main/java/seedu/planitarium/exceptions/UnknownException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package seedu.planitarium.exceptions;
 
 public class UnknownException extends Exception {
     protected static final String ERROR_MSG = "Sorry, please check your input again";
